requirements.txt
streamlit>=1.20.0
google-generativeai
protobuf==3.20.*
transformers>=4.21.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
